v1.1 imported (levenberg-marquardt minimize, new ui etc)

+++ ios/Group5iBeacons/Debug/Models/Beacon.h
@@ -19,7 +19,7 @@
 @property (strong, nonatomic) Location *location;
 
 - (instancetype)initWithMajor:(NSUInteger)major
-                        minor:(CGFloat)minor
+                        minor:(NSUInteger)minor
                      location:(Location *)location;
 
 - (BOOL)isEqualToBeacon:(Beacon *)beacon;
